backend/requirements.txt
fastapi
python-dotenv
pydantic
langchain-google-genai
uvicorn
requests
python-multipart
langchain